# Switch from internal 5-arg `searchsorted*` methods to views