--- a/src/linalg.jl
+++ b/src/linalg.jl
@@ -453,7 +453,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, tf
                     i1 = ia[j]
                     i2 = ia[j + 1] - 1
                     done = unit
-        
+
                     bj = B[joff + j]
                     for ii = i1:i2
                         jai = ja[ii]
@@ -483,7 +483,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, tf
                     i2 = ia[j + 1] - 1
                     akku = Z
                     j0 = !unit ? j : j - 1
-        
+
                     # loop through column j of A - only structural non-zeros
                     for ii = i1:i2
                         jai = ja[ii]
@@ -509,7 +509,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, tf
                     i1 = ia[j]
                     i2 = ia[j + 1] - 1
                     done = unit
-        
+
                     bj = B[joff + j]
                     for ii = i2:-1:i1
                         jai = ja[ii]
@@ -539,7 +539,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, tf
                     i2 = ia[j + 1] - 1
                     akku = Z
                     j0 = !unit ? j : j + 1
-        
+
                     # loop through column j of A - only structural non-zeros
                     for ii = i2:-1:i1
                         jai = ja[ii]
@@ -582,7 +582,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, ::
                 i1 = ia[j]
                 i2 = ia[j + 1] - 1
                 done = unit
-    
+
                 bj = B[joff + j]
                 for ii = i1:i2
                     jai = ja[ii]
@@ -610,7 +610,7 @@ function LinearAlgebra.generic_trimatmul!(C::StridedVecOrMat, uploc, isunitc, ::
                 i1 = ia[j]
                 i2 = ia[j + 1] - 1
                 done = unit
-    
+
                 bj = B[joff + j]
                 for ii = i2:-1:i1
                     jai = ja[ii]
@@ -664,7 +664,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, tf
                     i2 = ia[j + 1] - one(eltype(ia))
 
                     # find diagonal element
-                    ii = searchsortedfirst(ja, j, i1, i2, Base.Order.Forward)
+                    ii = searchsortedfirst(view(ja, i1:i2), j) - i1 + 1
                     jai = ii > i2 ? zero(eltype(ja)) : ja[ii]
 
                     cj = C[j,k]
@@ -693,7 +693,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, tf
                     i2 = ia[j + 1] - 1
                     akku = B[j,k]
                     done = false
-            
+
                     # loop through column j of A - only structural non-zeros
                     for ii = i2:-1:i1
                         jai = ja[ii]
@@ -721,11 +721,11 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, tf
                 for j = nrowB:-1:1
                     i1 = ia[j]
                     i2 = ia[j + 1] - one(eltype(ia))
-            
+
                     # find diagonal element
-                    ii = searchsortedlast(ja, j, i1, i2, Base.Order.Forward)
+                    ii = searchsortedlast(view(ja, i1:i2), j) - i1 + 1
                     jai = ii < i1 ? zero(eltype(ja)) : ja[ii]
-            
+
                     cj = C[j,k]
                     # check for zero pivot and divide with pivot
                     if jai == j
@@ -737,7 +737,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, tf
                     elseif !unit
                         throw(LinearAlgebra.SingularException(j))
                     end
-            
+
                     # update remaining part
                     for i = ii:-1:i1
                         C[ja[i],k] -= cj * LinearAlgebra._ustrip(aa[i])
@@ -752,7 +752,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, tf
                     i2 = ia[j + 1] - 1
                     akku = B[j,k]
                     done = false
-            
+
                     # loop through column j of A - only structural non-zeros
                     for ii = i1:i2
                         jai = ja[ii]
@@ -801,7 +801,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, ::
                 i2 = ia[j + 1] - one(eltype(ia))
 
                 # find diagonal element
-                ii = searchsortedfirst(ja, j, i1, i2, Base.Order.Forward)
+                ii = searchsortedfirst(view(ja, i1:i2), j) - i1 + 1
                 jai = ii > i2 ? zero(eltype(ja)) : ja[ii]
 
                 cj = C[j,k]
@@ -828,11 +828,11 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, ::
             for j = nrowB:-1:1
                 i1 = ia[j]
                 i2 = ia[j + 1] - one(eltype(ia))
-        
+
                 # find diagonal element
-                ii = searchsortedlast(ja, j, i1, i2, Base.Order.Forward)
+                ii = searchsortedlast(view(ja, i1:i2), j) - i1 + 1
                 jai = ii < i1 ? zero(eltype(ja)) : ja[ii]
-        
+
                 cj = C[j,k]
                 # check for zero pivot and divide with pivot
                 if jai == j
@@ -844,7 +844,7 @@ function LinearAlgebra.generic_trimatdiv!(C::StridedVecOrMat, uploc, isunitc, ::
                 elseif !unit
                     throw(LinearAlgebra.SingularException(j))
                 end
-        
+
                 # update remaining part
                 for i = ii:-1:i1
                     C[ja[i],k] -= cj * LinearAlgebra._ustrip(conj(aa[i]))
@@ -904,13 +904,13 @@ end
 function nzrangeup(A, i, excl=false)
     r = nzrange(A, i); r1 = r.start; r2 = r.stop
     rv = rowvals(A)
-    @inbounds r2 < r1 || rv[r2] <= i - excl ? r : r1:searchsortedlast(rv, i - excl, r1, r2, Forward)
+    @inbounds r2 < r1 || rv[r2] <= i - excl ? r : r1:(searchsortedlast(view(rv, r1:r2), i - excl) + r1-1)
 end
 # row range from diagonal (included if excl=false) to end
 function nzrangelo(A, i, excl=false)
     r = nzrange(A, i); r1 = r.start; r2 = r.stop
     rv = rowvals(A)
-    @inbounds r2 < r1 || rv[r1] >= i + excl ? r : searchsortedfirst(rv, i + excl, r1, r2, Forward):r2
+    @inbounds r2 < r1 || rv[r1] >= i + excl ? r : (searchsortedfirst(view(rv, r1:r2), i + excl) + r1-1):r2
 end
 
 dot(x::AbstractVector, A::RealHermSymComplexHerm{<:Any,<:AbstractSparseMatrixCSC}, y::AbstractVector) =
@@ -985,7 +985,7 @@ function _dot(x::SparseVector, A::AbstractSparseMatrixCSC, y::SparseVector, rang
         r1 = Int(Acolptr[i])
         r2 = Int(Acolptr[i+1]-1)
         r1 > r2 && continue
-        r1 = searchsortedfirst(Arowval, i, r1, r2, Forward)
+        r1 += searchsortedfirst(view(Arowval, r1:r2), i) - 1
         ((r1 > r2) || (Arowval[r1] != i)) && continue
         r += dot(x[i], diagop(Anzval[r1]), y[i])
     end