There should be little performance impact:
```julia
julia> x = cumsum(rand(1000));

julia> @btime searchsortedfirst($x, 250, 400, 600, Base.Forward)
  14.612 ns (0 allocations: 0 bytes)
510

julia> @btime searchsortedfirst(view($x, 400:600), 250)
  14.612 ns (0 allocations: 0 bytes)
111
```

See JuliaLang/julia#51177 and JuliaLang/julia#51176 for motivation