@fredrikekre 2023-09-10
The benchmark in the OP is not benchmarking the direct replacement. Is there a benchmark for e.g. `getindex(::SparseMatrixCSC, ::Int, ::Int)`?